Move fi_datatype_size into common sources.
THe sockets provider was using this and it broke the =dl build, also remove the identical copy that psm provider was maintaining.
